% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_time.R
\name{parse_time_hytek}
\alias{parse_time_hytek}
\title{Parse hytek time}
\usage{
parse_time_hytek(time)
}
\arguments{
\item{time}{a character vector containing the time scraped from Hy-Tek}
}
\description{
Parse hytek time
}
\examples{
parse_time_hytek(c("4:23.10#", "23.10"))
}
